requirements.txt
asttokens==2.4.1
certifi==2024.8.30
colorama==0.4.6
comm==0.2.2
debugpy==1.8.6
decorator==5.1.1
dill==0.3.9
dnspython==2.7.0
executing==2.1.0
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
Pygments==2.18.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==307
PyYAML==6.0.2
pyzmq==26.2.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.2
wcwidth==0.2.13
wheel==0.44.0

-e .